Splitting out fields and forms
