+++ edx_rbac/admin/forms.py
@@ -5,24 +5,9 @@ Forms to be used for djangoapps extending edx_rbac.
 from __future__ import absolute_import, unicode_literals
 
 from django import forms
-from django.contrib.auth.models import User
-from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext as _
 
-
-class UserFromEmailField(forms.EmailField):
-    """
-    Custom Form Field class for selecting users by entering an email.
-    Meant to be used for models with foreign keys to the user table.
-    """
-
-    def clean(self, value):
-        try:
-            user = User.objects.get(email=value)
-        except User.DoesNotExist:
-            raise ValidationError('User with email {} does not exist'.format(value))
-
-        return user
+from edx_rbac.fields import UserFromEmailField
 
 
 class UserRoleAssignmentAdminForm(forms.ModelForm):
